tests ported to pytest
