--- pyconcepticon/tests/fixtures/conceptlist.tsv
@@ -1,2 +1,0 @@
-ID	CONCEPTICON_ID	NUMBER	ENGLISH
-CL-2010-1-1	1732	1	sky
