--- pyconcepticon/tests/fixtures/conceptlist2.tsv
@@ -1,3 +1,0 @@
-ID	CONCEPTICON_ID	NUMBER	ENGLISH	CONCEPTICON_GLOSS
-CL-2-1	1732	1	sky	
-CL-2-2	1732	2	sky	DUST
