+++ requirements.txt
@@ -1,0 +1,3 @@
+# development environment: install in development mode
+-e .[dev,test]
+
