+++ setup.cfg
@@ -6,10 +6,17 @@ exclude = .tox
 [easy_install]
 zip_ok = false
 
-[nosetests]
-match=^test
-tests=pyconcepticon
-nocapture=1
-with-coverage=1
-cover-package=pyconcepticon
-cover-erase=1
+[tool:pytest]
+minversion = 3.1
+testpaths = tests
+addopts =
+    --cov
+    --cov-report=term
+    --cov-report=html
+mock_use_standalone_module = true
+
+[coverage:run]
+source = pyconcepticon
+
+[coverage:report]
+show_missing = true
